% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_panel.R
\name{add_panel}
\alias{add_panel}
\title{Add Panel dynamically}
\usage{
add_panel(panel, ..., session = getDefaultReactiveDomain())
}
\arguments{
\item{panel}{A panel object}

\item{...}{Other options passed to the API.
See \url{https://dockview.dev/docs/api/dockview/panelApi}.}
}
\description{
Add Panel dynamically
}
