% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugins.R
\name{default_remove_tab_callback}
\alias{default_remove_tab_callback}
\title{Default remove tab callback}
\usage{
default_remove_tab_callback()
}
\description{
An example of a JavaScript function that can be used as a default
when removing a tab/panel.
}
