% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\docType{package}
\name{dockViewR-package}
\alias{dockViewR}
\alias{dockViewR-package}
\title{dockViewR: Layout Manager Widget for R and Shiny}
\description{
Provides R bindings to the 'dockview' JavaScript library. Create rearrangeable grid layouts in seconds.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/cynkra/dockViewR}
  \item \url{https://cynkra.github.io/dockViewR/}
  \item Report bugs at \url{https://github.com/cynkra/dockViewR/issues}
}

}
\author{
\strong{Maintainer}: David Granjon \email{dgranjon@ymail.com}

Authors:
\itemize{
  \item Nelson Stevens \email{nelson.stevens92@gmail.com}
  \item Nicolas Bennett \email{nicolas@cynkra.com}
}

Other contributors:
\itemize{
  \item mathuo \email{github.mathuo@gmail.com} [copyright holder]
}

}
\keyword{internal}
