% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxy.R
\name{find_panel_group}
\alias{find_panel_group}
\title{Find a panel in a group mapping}
\usage{
find_panel_group(id, panel_groups_map)
}
\arguments{
\item{id}{Panel ID to search for.}

\item{panel_groups_map}{Returned by \code{\link[=get_panel_group]{get_panel_group()}}.}
}
\value{
A character string representing the group ID
containing the panel, or NULL if not found.
}
\description{
Given a panel ID and a mapping of groups to panels,
returns the group ID that contains the specified panel.
}
