% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_view.R
\name{grid_view-shiny}
\alias{grid_view-shiny}
\alias{gridViewOutput}
\alias{grid_view_output}
\alias{renderGridView}
\alias{render_grid_view}
\title{Shiny bindings for grid_view}
\usage{
gridViewOutput(outputId, width = "100\%", height = "400px")

grid_view_output(outputId, width = "100\%", height = "400px")

renderGridView(expr, env = parent.frame(), quoted = FALSE)

render_grid_view(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{output variable to read from}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{expr}{An expression that generates a grid_view}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}
}
\value{
\code{gridViewOutput} and grid_view_output\verb{return a Shiny output function that can be used in the UI definition. \code{renderGridView} and}render_grid_view\verb{return a Shiny render function that can be used in the server definition to render a}grid_view` element.
}
\description{
Output and render functions for using grid_view within Shiny
applications and interactive Rmd documents.
}
