% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_view.R
\name{grid_view}
\alias{grid_view}
\title{Create a grid view widget}
\usage{
grid_view(
  panels,
  ...,
  theme = c("light-spaced", "light", "abyss", "abyss-spaced", "dark", "vs", "dracula",
    "replit"),
  orientation = c("horizontal", "vertical"),
  hide_borders = FALSE,
  proportional_layout = FALSE,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{panels}{Widget configuration. Slot for \link{panel}.}

\item{...}{Other options. See
\url{https://dockview.dev/docs/api/dockview/options/}.}

\item{theme}{Theme. One of
\code{c("abyss", "dark", "light", "vs", "dracula", "replit")}.}

\item{orientation}{Grid orientation. Either \code{horizontal} or \code{vertical}.}

\item{hide_borders}{Whether to hide panel borders. Default to FALSE.}

\item{proportional_layout}{TBD.}

\item{width}{Widget width.}

\item{height}{Widget height.}

\item{elementId}{When used outside Shiny.}
}
\value{
An HTML widget object.
}
\description{
Creates an interactive grid view widget that enables flexible
layout management with resizable panels.
This is a wrapper around the dockview.dev
JavaScript library, providing a powerful interface for
creating IDE-like layouts in Shiny applications or R Markdown documents.
grid view is a subset of dock view (with some feature removed like drag and drop).
}
