% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugins.R
\name{is_add_tab_plugin}
\alias{is_add_tab_plugin}
\title{Check if object is an add tab plugin}
\usage{
is_add_tab_plugin(x)
}
\arguments{
\item{x}{An object to test.}
}
\value{
Logical value.
}
\description{
Check if object is an add tab plugin
}
