% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugins.R
\name{is_dock_view_plugin}
\alias{is_dock_view_plugin}
\title{Check if object is a dock view plugin of specific type}
\usage{
is_dock_view_plugin(x, type)
}
\arguments{
\item{x}{An object to test.}

\item{type}{Character string specifying plugin type ("add_tab" or "remove_tab").}
}
\value{
Logical value indicating whether the object is the specified plugin type.
}
\description{
Check if object is a dock view plugin of specific type
}
