% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge_group.R
\name{is_edge_group}
\alias{is_edge_group}
\title{Check whether an object is an edge group}
\usage{
is_edge_group(x)
}
\arguments{
\item{x}{An object to test.}
}
\value{
Logical value indicating whether \code{x} was created with \code{\link[=edge_group]{edge_group()}}.
}
\description{
Check whether an object is an edge group
}
