% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{new_dock_layout}
\alias{new_dock_layout}
\alias{is_dock_layout}
\alias{validate_dock_layout}
\alias{as_dock_layout}
\alias{layout_panel_ids}
\alias{empty_dock_layout}
\alias{tree_to_dock_layout}
\title{Dockview Layout Constructor}
\usage{
new_dock_layout(grid = NULL, panels = NULL, active_group = NULL)

is_dock_layout(x)

validate_dock_layout(x, ...)

as_dock_layout(x, ...)

layout_panel_ids(x)

empty_dock_layout()

tree_to_dock_layout(tree, remove, active_group = NULL)
}
\arguments{
\item{grid, panels, active_group}{Layout components}

\item{tree}{nested List representing the layout tree.}

\item{remove}{Character vector of panel IDs to mark as to be closable.}

\item{activeGroup}{Optional active group ID.}
}
\description{
Create a new \code{dock_layout} object that holds
a compatible layout for dockview.
}
