% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_group.R
\name{move_group}
\alias{move_group}
\title{Move Group dynamically}
\usage{
move_group(
  proxy,
  id,
  group,
  position = NULL,
  session = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{proxy}{Result of \code{\link[=dock_view]{dock_view()}} or a character with the ID of the dockview.}

\item{id}{Panel id.}

\item{group}{ID of the panel you want to move the target to. They must belong
to different groups.}

\item{position}{Panel position options: one of \code{"left", "right", "top", "bottom", "center"}.}

\item{session}{shiny session object.
See \url{https://dockview.dev/docs/api/dockview/panelApi}.}
}
\description{
Move Group dynamically
}
