% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_group2.R
\name{move_group2}
\alias{move_group2}
\title{Move a group dynamically}
\usage{
move_group2(
  proxy,
  source,
  destination,
  position = NULL,
  session = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{proxy}{Result of \code{\link[=dock_view]{dock_view()}} or a character with the ID of the dockview.}

\item{source}{Panel-id of a panel within the group that should be moved.}

\item{destination}{Panel-id of a panel within the group you want as a destination.}

\item{position}{Panel position options: one of \code{"left", "right", "top", "bottom", "center"}.}

\item{session}{shiny session object.
See \url{https://dockview.dev/docs/api/dockview/panelApi}.}
}
\description{
move_group2 moves a group to a different position from withing a shiny server function.
The parameter source refers to a panel-id of a panel within the group you want to move.
Likewise destination refers to a panel-id of a panel within the group you want to
select as destination.
The difference between move_group2 and move_group is that move_group2 selects both
source and destination by panel-id, whereas move_group selects by group-id.
}
