% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_panel.R
\name{move_panel}
\alias{move_panel}
\title{Move Panel dynamically}
\usage{
move_panel(
  dock,
  id,
  position = NULL,
  group = NULL,
  index = NULL,
  session = getDefaultReactiveDomain()
)
}
\arguments{
\item{dock}{Dock proxy object created with \code{\link[=dock_view_proxy]{dock_view_proxy()}}.}

\item{id}{Panel id.}

\item{position}{Panel position options: one of
\code{"left", "right", "top", "bottom", "center"}.}

\item{group}{ID of a panel that belongs to the target group. The moving panel
and this panel must belong to different groups.}

\item{index}{Panel index. If panels belong to the same group,
you can use index to move the target panel at the desired position.
When group is left NULL, index must be passed and cannot exceed the
total number of panels or be negative.}

\item{session}{shiny session object.
See \url{https://dockview.dev/docs/api/dockview/panelApi/}.}
}
\value{
The dock proxy object, invisibly.
}
\description{
Move Panel dynamically
}
