% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dock_view.R
\name{panel}
\alias{panel}
\title{Dock panel}
\usage{
panel(id, title, content, active = TRUE, ...)
}
\arguments{
\item{id}{Panel unique id.}

\item{title}{Panel title.}

\item{content}{Panel content. Can be a list of Shiny tags.}

\item{active}{Is active?}

\item{...}{Other options passed to the API.
See \url{https://dockview.dev/docs/api/dockview/panelApi}.}
}
\description{
Create a dock panel
}
