% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_panel.R
\name{remove_panel}
\alias{remove_panel}
\title{Remove Panel dynamically}
\usage{
remove_panel(dock, id, session = getDefaultReactiveDomain())
}
\arguments{
\item{dock}{Dock proxy object created with \code{\link[=dock_view_proxy]{dock_view_proxy()}}.}

\item{id}{Id of the panel that ought to be removed.}

\item{session}{shiny session object.
See \url{https://dockview.dev/docs/api/dockview/panelApi/}.}
}
\value{
The dock proxy object, invisibly.
}
\description{
Remove Panel dynamically
}
