% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_panel.R
\name{select_panel}
\alias{select_panel}
\title{Select a panel dynamically}
\usage{
select_panel(dock, id, session = getDefaultReactiveDomain())
}
\arguments{
\item{dock}{Dock proxy object created with \code{\link[=dock_view_proxy]{dock_view_proxy()}}.}

\item{id}{Panel id.}

\item{session}{shiny session object.
See \url{https://dockview.dev/docs/api/dockview/panelApi/}.}
}
\value{
The dock proxy object, invisibly.
}
\description{
Select a panel dynamically
}
