% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_renderer.R
\name{set_panel_renderer}
\alias{set_panel_renderer}
\title{Select a panel renderer dynamically}
\usage{
set_panel_renderer(
  dock_id,
  id,
  renderer = c("always", "onlyWhenVisible"),
  session = getDefaultReactiveDomain()
)
}
\arguments{
\item{dock_id}{Dock unique id. When using modules the namespace
is automatically added.}

\item{id}{Panel id.}

\item{renderer}{Renderer value. Either \code{"always"} or \code{"onlyWhenVisible"}.}

\item{session}{shiny session object.
See \url{https://dockview.dev/docs/api/dockview/panelApi/}.}
}
\value{
This function is called for its side effect.
It sends a message to JavaScript through the current websocket connection,
leveraging the shiny session object.
}
\description{
Select a panel renderer dynamically
}
