% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_panel_size.R
\name{set_panel_size}
\alias{set_panel_size}
\title{Change panel size dynamically}
\usage{
set_panel_size(dock_id, id, height, session = getDefaultReactiveDomain())
}
\arguments{
\item{dock_id}{Dock unique id. When using modules the namespace
is automatically added.}

\item{id}{Panel id.}

\item{height}{New height of the panel.}

\item{session}{shiny session object.
See \url{https://dockview.dev/docs/core/panels/resizing}.}
}
\value{
This function is called for its side effect.
It sends a message to JavaScript through the current websocket connection,
leveraging the shiny session object.
}
\description{
Change panel size dynamically
}
