% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_panel_title.R
\name{set_panel_title}
\alias{set_panel_title}
\title{Set a panel title dynamically}
\usage{
set_panel_title(dock_id, id, title, session = getDefaultReactiveDomain())
}
\arguments{
\item{dock_id}{Dock unique id. When using modules the namespace
is automatically added.}

\item{id}{Panel id.}

\item{title}{New panel title.}

\item{session}{shiny session object.
See \url{https://dockview.dev/docs/api/dockview/panelApi/}.}
}
\value{
This function is called for its side effect.
It sends a message to JavaScript through the current websocket connection,
leveraging the shiny session object.
}
\description{
Set a panel title dynamically
}
