% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dock_view.R
\name{update_dock_view}
\alias{update_dock_view}
\title{Update options for dockview instance}
\usage{
update_dock_view(options, session = getDefaultReactiveDomain())
}
\arguments{
\item{options}{List of options for the \link{dock_view} instance.}

\item{session}{Shiny session object.}
}
\description{
This does not rerender the widget, just update options like global theme.
}
