% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugins.R
\name{validate_plugins}
\alias{validate_plugins}
\alias{validate_dock_view_plugin}
\alias{validate_common_plugin_fields}
\alias{validate_dock_view_plugin.dock_view_plugin_add_tab}
\alias{validate_dock_view_plugin.dock_view_plugin_remove_tab}
\title{Validate dock view plugins}
\usage{
validate_dock_view_plugin(plugin)

validate_common_plugin_fields(plugin)

\method{validate_dock_view_plugin}{dock_view_plugin_add_tab}(plugin)

\method{validate_dock_view_plugin}{dock_view_plugin_remove_tab}(plugin)
}
\arguments{
\item{plugin}{A dock view plugin object.}
}
\value{
The validated plugin object.
}
\description{
Internal validation functions for dock view plugins.
}
\keyword{internal}
