% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_panel.R
\name{add_panel}
\alias{add_panel}
\title{Add Panel dynamically}
\usage{
add_panel(dock, panel, ..., session = getDefaultReactiveDomain())
}
\arguments{
\item{dock}{Dock proxy object created with \code{\link[=dock_view_proxy]{dock_view_proxy()}}.}

\item{panel}{A panel object. See \link{panel} for the different parameters.}

\item{...}{Other options passed to the API. Not used yet.}

\item{session}{shiny session object.
See \url{https://dockview.dev/docs/api/dockview/panelApi/}.}
}
\value{
The dock proxy object, invisibly.
}
\description{
Add Panel dynamically
}
\seealso{
\code{\link[=panel]{panel()}}
}
