% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugins.R
\name{default_add_tab_callback}
\alias{default_add_tab_callback}
\title{Default add tab callback}
\usage{
default_add_tab_callback()
}
\description{
An example of a JavaScript function that can be used as a default
when adding a new tab/panel.
}
