% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dock_state.R
\name{get_dock}
\alias{get_dock}
\alias{get_panels}
\alias{get_panels_ids}
\alias{get_active_group}
\alias{get_grid}
\alias{get_groups}
\alias{get_groups_ids}
\alias{get_groups_panels}
\title{get dock}
\usage{
get_dock(proxy, session = getDefaultReactiveDomain())

get_panels(proxy, session = getDefaultReactiveDomain())

get_panels_ids(proxy, session = getDefaultReactiveDomain())

get_active_group(proxy, session = getDefaultReactiveDomain())

get_grid(proxy, session = getDefaultReactiveDomain())

get_groups(proxy, session = getDefaultReactiveDomain())

get_groups_ids(proxy, session = getDefaultReactiveDomain())

get_groups_panels(proxy, session = getDefaultReactiveDomain())
}
\arguments{
\item{proxy}{Result of \code{\link[=dock_view]{dock_view()}} or a character with the ID of the dockview.}

\item{session}{shiny session object.}
}
\description{
get dock

get dock panels

get dock panels ids

get dock active group

get dock grid

get dock groups

get dock groups ids

get dock groups panels
}
\note{
Only works with server side functions like \link{add_panel}. Don't call it
from the UI.
}
