% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dock_view.R
\name{dock_view}
\alias{dock_view}
\title{Dock view widget}
\usage{
dock_view(message, width = NULL, height = NULL, elementId = NULL)
}
\arguments{
\item{message}{Widget configuration.}

\item{width}{Widget width.}

\item{height}{Widget height.}

\item{elementId}{When used outside Shiny.}
}
\description{
Create a dock view widget
}
