% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dock_view.R
\name{dock_view}
\alias{dock_view}
\title{Create a dock view widget}
\usage{
dock_view(
  panels,
  ...,
  theme = c("light-spaced", "light", "abyss", "abyss-spaced", "dark", "vs", "dracula",
    "replit"),
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{panels}{Widget configuration. Slot for \link{panel}.}

\item{...}{Other options. See \url{https://dockview.dev/docs/api/dockview/options}.}

\item{theme}{Theme. One of
\code{c("abyss", "dark", "light", "vs", "dracula", "replit")}.}

\item{width}{Widget width.}

\item{height}{Widget height.}

\item{elementId}{When used outside Shiny.}
}
\value{
An HTML widget object.
}
\description{
Create a dock view widget for user controled layout management
and docking controls.
}
\section{Examples in Shinylive}{
\describe{
  \item{example-1}{
    \href{https://shinylive.io/r/app/#code=NobwRAdghgtgpmAXGKAHVA6ASmANGAYwHsIAXOMpMAdzgCMAnRRASwgGdSoAbbgCgA6YACZECAawBqLONSxDcAAgZxURdooC8ioQAtSpVO0QB6EwzbkGAMygE4GAOYtSugK50MLIuctwbdnAAtNRQ7DBBBAxQECZCAJQCENwsjFAMAJ587LpsGYnJqdGZfHTsKXQFFcVZohLSslgFSW4sigA8QYqoUI5wAPrWLLxQdNxwghCKinVSMtQA8m6Gy4IiYuIJSc0c-gBu-h1d1m4QBKTeEHxsqMtKRMu3pErscOzsl-GKIEmKDyukAAksyOygown8ABENg1qHwflNprN+nt5pNphiehA4NwNNoUpx0RjMTEcUTiRiWMItDowABGBS-CkYi6kcY0oQABVJ3EUDLwTOZimIZAopBpXEcABkWITBULFOUqf4AJIQJ7khXTIREMqMxFa7VgABybhgdEORGsfzK+ygFxIxn1hoxMDYNIADLh5VqYFAAB40ukekPeg2GvY8NxwGkAVhDPuZ8TDLtQ3CIpCWALWwllpE56dIW3DFIKQuTie6PM1FKpHLAACZnQrWeztFyeYomwKS8SReQyBLejK5b2Ka9xuc1RrK8yhJGLKNxs2XUJJOkWEu4IgV4aCGsAMIZFIQCEMdhCesEY8KWkAFWiHDd70ul-bKBgt6EAHE4OkL2A9Z9OkxYuhWY7ElwYxwFmM4iPaUCgQq4FamoHwOlM+J5jWzIqNY-gUPY3LYry778imhq5io5yXPWFiOPoQizooZZJj6KGQbocDwHRqgpEWYCCmWAC+Ox-I8yzAnmBYZqCKinv4MmkPCgoqAAjtc6qSbq7CsYouSEgwEBEAwMCaU8gI6fEIllv8FnCAhcngv4d5buSCIUjApAEP+wBKPuQgwKgji3jckkLpu0HJjMDBEKgNIAGIAIJSgAygAogAuoKwkUcoRDUNA8B4ood5YAAqulvwFMJSQ5HkSXoHwrQvPs-jxGAwmZUAA}{Open in Shinylive}
    \if{html}{\out{<iframe class="iframe_shinylive" src="https://shinylive.io/r/app/#code=NobwRAdghgtgpmAXGKAHVA6ASmANGAYwHsIAXOMpMAdzgCMAnRRASwgGdSoAbbgCgA6YACZECAawBqLONSxDcAAgZxURdooC8ioQAtSpVO0QB6EwzbkGAMygE4GAOYtSugK50MLIuctwbdnAAtNRQ7DBBBAxQECZCAJQCENwsjFAMAJ587LpsGYnJqdGZfHTsKXQFFcVZohLSslgFSW4sigA8QYqoUI5wAPrWLLxQdNxwghCKinVSMtQA8m6Gy4IiYuIJSc0c-gBu-h1d1m4QBKTeEHxsqMtKRMu3pErscOzsl-GKIEmKDyukAAksyOygown8ABENg1qHwflNprN+nt5pNphiehA4NwNNoUpx0RjMTEcUTiRiWMItDowABGBS-CkYi6kcY0oQABVJ3EUDLwTOZimIZAopBpXEcABkWITBULFOUqf4AJIQJ7khXTIREMqMxFa7VgABybhgdEORGsfzK+ygFxIxn1hoxMDYNIADLh5VqYFAAB40ukekPeg2GvY8NxwGkAVhDPuZ8TDLtQ3CIpCWALWwllpE56dIW3DFIKQuTie6PM1FKpHLAACZnQrWeztFyeYomwKS8SReQyBLejK5b2Ka9xuc1RrK8yhJGLKNxs2XUJJOkWEu4IgV4aCGsAMIZFIQCEMdhCesEY8KWkAFWiHDd70ul-bKBgt6EAHE4OkL2A9Z9OkxYuhWY7ElwYxwFmM4iPaUCgQq4FamoHwOlM+J5jWzIqNY-gUPY3LYry778imhq5io5yXPWFiOPoQizooZZJj6KGQbocDwHRqgpEWYCCmWAC+Ox-I8yzAnmBYZqCKinv4MmkPCgoqAAjtc6qSbq7CsYouSEgwEBEAwMCaU8gI6fEIllv8FnCAhcngv4d5buSCIUjApAEP+wBKPuQgwKgji3jckkLpu0HJjMDBEKgNIAGIAIJSgAygAogAuoKwkUcoRDUNA8B4ood5YAAqulvwFMJSQ5HkSXoHwrQvPs-jxGAwmZUAA" style="height: 800px; width: 100vw; max-width: 1400px; border: 1px solid rgba(0,0,0,0.175); border-radius: .375rem; position: absolute; left: 50\%; margin-top: 30px; transform: translateX(-50\%); z-index: 1"></iframe>}}
    \if{html}{\out{<a style='height: 800px; display: block;'></a>}}
  }
}
}

