% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dock_view.R
\name{dock_view}
\alias{dock_view}
\title{Dock view widget}
\usage{
dock_view(
  panels,
  ...,
  theme = c("light-spaced", "light", "abyss", "abyss-spaced", "dark", "vs", "dracula",
    "replit"),
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{panels}{Widget configuration. Slot for \link{panel}.}

\item{...}{Other options. See \url{https://dockview.dev/docs/api/dockview/options}.}

\item{theme}{Theme. One of
\code{c("abyss", "dark", "light", "vs", "dracula", "replit")}.}

\item{width}{Widget width.}

\item{height}{Widget height.}

\item{elementId}{When used outside Shiny.}
}
\description{
Create a dock view widget
}
\section{Examples in Shinylive}{
\describe{
  \item{example-1}{
    \href{https://shinylive.io/r/app/#code=NobwRAdghgtgpmAXGKAHVA6ASmANGAYwHsIAXOMpMAdzgCMAnRRASwgGdSoAbbgCgA6YACZECAawBqLONSxDcAAgZxURdooC8ioQAtSpVO0QB6EwzbkGAMygE4GAOYtSugK50MLIuctwbdnAAtNRQ7DBBBAxQECZCAJQCENwsjFAMAJ587LpsGYnJqdGZfHTsKXQFFcVZohLSslgFSW4sigA8QYqoUI5wAPrWLLxQdNxwghCKinXi-QBuMtQA8m6Ga4IiYuIJSc0c-vP+HV3WbhAEpN4QfGyoa0pEa-ekSuxw7OzX8YogSYpPdakAAksxOygown8ABFtgslnw-lNprN4bJJtNMT0IHBuBptClOBjMViYrjiSTMSxhFodGAAIwKf6UzFXUjjWlCAAKZO4ikZeGZLMUxDIFFItK4jgAMiwiULhYpytT-ABJCAvCmK6ZCIhlJnI7U6sAAOTcMDoxyI1gBZUOUCuJGMBqNmJgbFpAAZcArtTAoAAPWn0z2hn2Go3zHhuOC0gCsod9LPi4ddqG4RFIqyBm2EctIXIzpF2EcpBWFKaT3V5Wsp1M5YAATC7FWyOdpubzFM3BaWSaLyGRJb1ZfK+5T3uNLurNVWWUIoxZRuMW66hJJ0ixl3BEKujQRNgBhDIpCBQhjsIQNggnhR0gAq0Q47s+1yvHZQMDvQgA4nB0peYANn06Qlq6lbjiSXBjHA2aziIDpQGBioQdqahfI6UwEvmtYsio1j+BQ9g8jifIfgKqZGnmKiXNcDYWI4+hCHOijlsmvqoVBuhwPA9GqCkxZgEK5YAL77ACzxrKC+aFpm4IqGe-iyaQiJCioACOtwalJersGxii5ESDAQEQDAwFpLzArp8SieWgKWcIiHyZC-j3tuFJIpSMCkAQAHAEoB5CDAqCOHedxSYuW4wSmMwMEQqC0gAYgAgtKADKACiAC6QoiZRyhENQ0DwPiij3lgACqGX-AUIlJDkeTJegfCtG8hz+PEYAiVlQA}{Open in Shinylive}
    \if{html}{\out{<iframe class="iframe_shinylive" src="https://shinylive.io/r/app/#code=NobwRAdghgtgpmAXGKAHVA6ASmANGAYwHsIAXOMpMAdzgCMAnRRASwgGdSoAbbgCgA6YACZECAawBqLONSxDcAAgZxURdooC8ioQAtSpVO0QB6EwzbkGAMygE4GAOYtSugK50MLIuctwbdnAAtNRQ7DBBBAxQECZCAJQCENwsjFAMAJ587LpsGYnJqdGZfHTsKXQFFcVZohLSslgFSW4sigA8QYqoUI5wAPrWLLxQdNxwghCKinXi-QBuMtQA8m6Ga4IiYuIJSc0c-vP+HV3WbhAEpN4QfGyoa0pEa-ekSuxw7OzX8YogSYpPdakAAksxOygown8ABFtgslnw-lNprN4bJJtNMT0IHBuBptClOBjMViYrjiSTMSxhFodGAAIwKf6UzFXUjjWlCAAKZO4ikZeGZLMUxDIFFItK4jgAMiwiULhYpytT-ABJCAvCmK6ZCIhlJnI7U6sAAOTcMDoxyI1gBZUOUCuJGMBqNmJgbFpAAZcArtTAoAAPWn0z2hn2Go3zHhuOC0gCsod9LPi4ddqG4RFIqyBm2EctIXIzpF2EcpBWFKaT3V5Wsp1M5YAATC7FWyOdpubzFM3BaWSaLyGRJb1ZfK+5T3uNLurNVWWUIoxZRuMW66hJJ0ixl3BEKujQRNgBhDIpCBQhjsIQNggnhR0gAq0Q47s+1yvHZQMDvQgA4nB0peYANn06Qlq6lbjiSXBjHA2aziIDpQGBioQdqahfI6UwEvmtYsio1j+BQ9g8jifIfgKqZGnmKiXNcDYWI4+hCHOijlsmvqoVBuhwPA9GqCkxZgEK5YAL77ACzxrKC+aFpm4IqGe-iyaQiJCioACOtwalJersGxii5ESDAQEQDAwFpLzArp8SieWgKWcIiHyZC-j3tuFJIpSMCkAQAHAEoB5CDAqCOHedxSYuW4wSmMwMEQqC0gAYgAgtKADKACiAC6QoiZRyhENQ0DwPiij3lgACqGX-AUIlJDkeTJegfCtG8hz+PEYAiVlQA" style="height: 800px; width: 100vw; max-width: 1400px; border: 1px solid rgba(0,0,0,0.175); border-radius: .375rem; position: absolute; left: 50\%; margin-top: 30px; transform: translateX(-50\%); z-index: 1"></iframe>}}
    \if{html}{\out{<a style='height: 800px; display: block;'></a>}}
  }
}
}

