% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dock_view.R
\name{dock_view}
\alias{dock_view}
\title{Dock view widget}
\usage{
dock_view(
  panels,
  ...,
  theme = c("light", "abyss", "dark", "vs", "dracula", "replit"),
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{panels}{Widget configuration. Slot for \link{panel}.}

\item{...}{Other options. See \url{https://dockview.dev/docs/api/dockview/options}.}

\item{theme}{Theme. One of
\code{c("abyss", "dark", "light", "vs", "dracula", "replit")}.}

\item{width}{Widget width.}

\item{height}{Widget height.}

\item{elementId}{When used outside Shiny.}
}
\description{
Create a dock view widget
}
