% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\docType{package}
\name{dockViewR-package}
\alias{dockViewR}
\alias{dockViewR-package}
\title{dockViewR: Layout Manager Widget for R and 'shiny' Apps}
\description{
Provides R bindings to the 'dockview' 'JavaScript' library \url{https://dockview.dev/}. Create fully customizable grid layouts (docks) in seconds to include in interactive R reports with R Markdown or 'Quarto' or in 'shiny' apps \url{https://shiny.posit.co/}. In 'shiny' mode, modify docks by dynamically adding, removing or moving panels or groups of panels from the server function. Choose among 8 stunning themes (dark and light), serialise the state of a dock to restore it later.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/cynkra/dockViewR}
  \item \url{https://cynkra.github.io/dockViewR/}
  \item Report bugs at \url{https://github.com/cynkra/dockViewR/issues}
}

}
\author{
\strong{Maintainer}: David Granjon \email{dgranjon@ymail.com}

Authors:
\itemize{
  \item Nelson Stevens \email{nelson.stevens92@gmail.com}
  \item Nicolas Bennett \email{nicolas@cynkra.com}
}

Other contributors:
\itemize{
  \item mathuo \email{github.mathuo@gmail.com} [copyright holder]
  \item cynkra GmbH \email{mail@cynkra.com} [funder]
}

}
\keyword{internal}
