% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_panel.R
\name{move_panel}
\alias{move_panel}
\title{Move Panel dynamically}
\usage{
move_panel(
  proxy,
  id,
  position,
  group = NULL,
  index = NULL,
  session = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{proxy}{Result of \code{\link[=dock_view]{dock_view()}} or a character with the ID of the dockview.}

\item{id}{Panel id.}

\item{position}{Panel position options. A list with \code{referencePanel} and \code{direction}.}

\item{group}{Panel group.}

\item{index}{Panel index.}

\item{session}{shiny session object.
See \url{https://dockview.dev/docs/api/dockview/panelApi}.}
}
\description{
Move Panel dynamically
}
