% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_panel.R
\name{remove_panel}
\alias{remove_panel}
\title{Remove Panel dynamically}
\usage{
remove_panel(id, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{session}{shiny session object.
See \url{https://dockview.dev/docs/api/dockview/panelApi}.}

\item{proxy}{Result of \code{\link[=dock_view]{dock_view()}} or a character with the ID of the dockview.}

\item{panel}{A panel object}

\item{...}{Other options passed to the API.}
}
\description{
Remove Panel dynamically
}
