% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_panel.R
\name{remove_panel}
\alias{remove_panel}
\title{Remove Panel dynamically}
\usage{
remove_panel(proxy, id, session = getDefaultReactiveDomain())
}
\arguments{
\item{proxy}{Result of \code{\link[=dock_view]{dock_view()}} or a character with the ID of the dockview.}

\item{id}{Id of the panel that ought to be removed.}

\item{session}{shiny session object.
See \url{https://dockview.dev/docs/api/dockview/panelApi}.}
}
\description{
Remove Panel dynamically
}
