% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_panel.R
\name{remove_panel}
\alias{remove_panel}
\title{Remove Panel dynamically}
\usage{
remove_panel(dock_id, id, session = getDefaultReactiveDomain())
}
\arguments{
\item{dock_id}{Dock unique id. When using modules the namespace is
automatically added.}

\item{id}{Id of the panel that ought to be removed.}

\item{session}{shiny session object.
See \url{https://dockview.dev/docs/api/dockview/panelApi/}.}
}
\value{
This function is called for its side effect.
It sends a message to JavaScript through the current websocket connection,
leveraging the shiny session object.
}
\description{
Remove Panel dynamically
}
