% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_panel.R
\name{select_panel}
\alias{select_panel}
\title{Select a panel dynamically}
\usage{
select_panel(dock_id, id, session = getDefaultReactiveDomain())
}
\arguments{
\item{dock_id}{Dock unique id. When using modules the namespace
is automatically added.}

\item{id}{Panel id.}

\item{session}{shiny session object.
See \url{https://dockview.dev/docs/api/dockview/panelApi/}.}
}
\value{
This function is called for its side effect.
It sends a message to JavaScript through the current websocket connection,
leveraging the shiny session object.
}
\description{
Select a panel dynamically
}
