% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxy.R
\name{update_dock_view}
\alias{update_dock_view}
\title{Update options for dockview instance}
\usage{
update_dock_view(dock, options)
}
\arguments{
\item{dock}{Dock proxy created with \code{\link[=dock_view_proxy]{dock_view_proxy()}}.}

\item{options}{List of options for the \link{dock_view} instance.}
}
\value{
This function is called for its side effect.
It sends a message to JavaScript through the current websocket connection,
leveraging the shiny session object.
}
\description{
This does not rerender the widget, just update options like global theme.
}
